+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Replaced [deprecated request module](https://github.com/request/request#deprecated) with `got`.
+
 ## v3.1.0, 2020-03-09
 
 - Fixed express-sharp to not return always jpeg. (kriscarle, [#23](https://github.com/pmb0/express-sharp/pull/23))